backend/requirements.txt
# Core Backend Dependencies
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
boto3==1.34.69  # Required for AWS integration

# Data Processing
pandas==2.2.3
numpy==2.2.3
pandas_market_calendars==4.6.1
exchange_calendars==4.9
python-dateutil==2.9.0.post0
pytz==2025.1
simplejson==3.20.1
beautifulsoup4==4.12.3  # Required for web scraping

# Database
psycopg2-binary==2.9.10

# Testing
pytest==8.3.5
pytest-cov==6.0.0
coverage==7.6.12

# Configuration
python-dotenv==1.0.1

# Development Tools
chardet==5.2.0  # Required for file integrity checks